# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BoxMeOut API
  version: 0.2.0
  description: REST API for the BoxMeOut decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
...
paths:
...
  /auth/login:
    post:
      summary: Login with email and password
      operationId: login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Tokens issued, or tempToken when 2FA is enabled
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
                  tempToken:    { type: string, description: Present only when 2FA is required }
                  requires2FA:  { type: boolean }
        '400':
          description: Missing credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/forgot-password:
    post:
      summary: Request a password-reset email
      operationId: forgotPassword
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        '200':
          description: Reset link sent (always returned to prevent email enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
  /auth/reset-password:
    post:
      summary: Reset password using a reset token
      operationId: resetPassword
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:       { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/2fa/setup:
    post:
      summary: Generate TOTP secret and QR code
      operationId: setup2FA
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: TOTP secret and QR code
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:   { type: string }
                  qrCode:   { type: string, description: Base64 PNG data URI }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/2fa/enable:
    post:
      summary: Enable 2FA after verifying TOTP code
      operationId: enable2FA
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp]
              properties:
                otp: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: 2FA enabled
        '401':
          description: Unauthorized or invalid OTP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/2fa/disable:
    post:
      summary: Disable 2FA
      operationId: disable2FA
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp]
              properties:
                otp: { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: 2FA disabled
        '401':
          description: Unauthorized or invalid OTP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/2fa/verify:
    post:
      summary: Complete login by verifying 2FA OTP
      operationId: verify2FA
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tempToken, otp]
              properties:
                tempToken: { type: string }
                otp:       { type: string, minLength: 6, maxLength: 6 }
      responses:
        '200':
          description: Access and refresh tokens issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:  { type: string }
                  refreshToken: { type: string }
        '401':
          description: Invalid temp token or OTP
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ---------------------------------------------------------------------------
  # Users
  # ---------------------------------------------------------------------------
...